+++ packages/opencode/tsconfig.json
@@ -13,12 +13,6 @@
     "declaration": true,
     "noEmit": true
   },
-  "include": [
-    "src/**/*",
-    "__tests__/**/*",
-    "tsup.config.ts",
-    "vitest.config.ts",
-    "eslint.config.js"
-  ],
+  "include": ["src/**/*", "tsup.config.ts", "eslint.config.js"],
   "exclude": ["node_modules", "dist"]
 }
